Fixed issue with prune.py when running network generation using a bash script
Issue: TypeError: 'bool' object is not iterable

The instance of pruner class in the bash script code was feeding use_compendia (a boolean) as the third parameter in the pruner class.  The actual third parameter is phospho_type (a str), the code was assigning the value of use_compendia to phospho_type causing an error.